+++ apps/3000-home/next.config.js
@@ -1,5 +1,5 @@
 const { withNx } = require('@nrwl/next/plugins/with-nx');
-const {createDelegatedModule} = require('@module-federation/utilities');
+const { createDelegatedModule } = require('@module-federation/utilities');
 const NextFederationPlugin = require('@module-federation/nextjs-mf');
 
 /**
@@ -17,7 +17,9 @@ const nextConfig = {
 
     const remotes = {
       shop: createDelegatedModule(require.resolve('./remote-delegate.js'), {
-        remote: `shop@http://localhost:3001/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,
+        remote: `shop@http://localhost:3001/_next/static/${
+          isServer ? 'ssr' : 'chunks'
+        }/remoteEntry.js`,
       }),
       // checkout: createDelegatedModule(require.resolve('./remote-delegate.js'), {
       //   remote: `checkout@http://localhost:3002/_next/static/${isServer ? 'ssr' : 'chunks'}/remoteEntry.js`,