+++ apps/3002-checkout/remote-delegate.js
@@ -11,12 +11,16 @@ module.exports = new Promise((resolve, reject) => {
 
   importDelegatedModule({
     global,
-    url: url + '?' + Date.now()
+    url: url + '?' + Date.now(),
   })
     .then(async (remote) => {
-      console.log(__resourceQuery, 'resolved remote from', __webpack_runtime_id__)
+      console.log(
+        __resourceQuery,
+        'resolved remote from',
+        __webpack_runtime_id__
+      );
 
-      resolve(remote)
+      resolve(remote);
     })
     .catch((err) => reject(err));
 });